+++ familyhelper-project-impl/src/main/java/com/dwarfeng/familyhelper/project/impl/util/FtpConstants.java
@@ -1,0 +1,16 @@
+package com.dwarfeng.familyhelper.project.impl.util;
+
+/**
+ * FTP 常量。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public final class FtpConstants {
+
+    public static final String PATH_MEMO_FILE = "familyhelper-project/memo-file";
+
+    private FtpConstants() {
+        throw new IllegalStateException("禁止实例化");
+    }
+}