+++ familyhelper-project-stack/src/main/java/com/dwarfeng/familyhelper/project/stack/bean/dto/MemoFile.java
@@ -1,0 +1,51 @@
+package com.dwarfeng.familyhelper.project.stack.bean.dto;
+
+import com.dwarfeng.subgrade.stack.bean.dto.Dto;
+
+import java.util.Arrays;
+
+/**
+ * 备忘录文件。
+ *
+ * @author DwArFeng
+ * @since 1.0.0
+ */
+public class MemoFile implements Dto {
+
+    private static final long serialVersionUID = 5578398856675359080L;
+
+    private String originName;
+    private byte[] content;
+
+    public MemoFile() {
+    }
+
+    public MemoFile(String originName, byte[] content) {
+        this.originName = originName;
+        this.content = content;
+    }
+
+    public String getOriginName() {
+        return originName;
+    }
+
+    public void setOriginName(String originName) {
+        this.originName = originName;
+    }
+
+    public byte[] getContent() {
+        return content;
+    }
+
+    public void setContent(byte[] content) {
+        this.content = content;
+    }
+
+    @Override
+    public String toString() {
+        return "MemoFile{" +
+                "originName='" + originName + '\'' +
+                ", content=" + Arrays.toString(content) +
+                '}';
+    }
+}